requirements.txt
confluent-kafka>=2.3.0
protobuf>=4.25.0
boto3>=1.34.0
pyyaml>=6.0
aws-embedded-metrics>=2.1.0
